Move input handling to main.py
